Ensuring the code generated for an enum and record state machine are similar
This does have the caveat of producing empty or "useless" methods.
The tradeoff is that I have the ability to call these methods
unconditionally in generated code which really helps with understanding
the code that generates code. I'm hoping the java compiler is smart
enough to compile out the extra methods but worst case they should
be pretty fast.